+++ src/editor/InspectorPanel.h
@@ -16,5 +16,6 @@ private:
     Scene *m_scene;
     char m_tag_buffer[256] = {};
     char m_mesh_buffer[256] = {};
+    char m_script_buffer[256] = {};
     int m_last_selected_id = -1;
 };